app/requirements/requirements.txt
asgiref==3.6.0
attrs==22.2.0
cfgv==3.3.1
coverage==7.2.1
distlib==0.3.6
Django==4.1.7
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==17.6.0
filelock==3.9.0
gunicorn==20.1.0
identify==2.5.19
inflection==0.5.1
iniconfig==2.0.0
nodeenv==1.7.0
packaging==23.0
platformdirs==3.1.0
pluggy==1.0.0
pre-commit==3.1.1
psycopg2-binary==2.9.5
pytest==7.2.2
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
virtualenv==20.20.0
